# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Links API",
        "description": "...",
        "version": "1.0.0"
    },
    "paths": {
...
        "/links": {
            "get": {
                "tags": [
                    "Links"
                ],
                "summary": "Get paginated list of links.",
                "operationId": "b94ffb5979c83b6ffd667b522d898802",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page of links",
                        "required": false,
                        "example": "page=1"
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Sorting order",
                        "required": false,
                        "example": "orderBy=id"
                    },
                    {
                        "name": "dir",
                        "in": "query",
                        "description": "Sorting direction",
                        "required": false,
                        "example": "dir=desc"
                    },
                    {
                        "name": "groups",
                        "in": "query",
                        "description": "Display links filtered by groups",
                        "required": false,
                        "example": "groups[]=1&groups[]=2"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "example": "search=Yandex"
                    },
                    {
                        "name": "searchBy",
                        "in": "query",
                        "description": "Column for searching",
                        "required": false,
                        "example": "searchBy=name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LinksListPaginated"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Permissions error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Links"
                ],
                "summary": "Create new link.",
                "description": "Remark: If you do not specify the groups, the default one will be automatically set.",
                "operationId": "0a6b0569be350f4a81e51be1888e3bf5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreLinkRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paginated list.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Info message."
                                        },
                                        "entity": {
                                            "$ref": "#/components/schemas/Link"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Permissions error."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/links/{id}": {
            "get": {
                "tags": [
                    "Links"
                ],
                "summary": "Show link by id.",
                "operationId": "068417c3fc1ff620108c70ba259b66be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of link.",
                        "required": true,
                        "example": "1"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Type of statistic sort (asc | desc).",
                        "example": "asc"
                    },
                    {
                        "name": "by",
                        "in": "query",
                        "description": "Sorting statistic by: (day | month | year | hits). Nothing = deafult = full date.",
                        "example": "day"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Link object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "entity": {
                                            "$ref": "#/components/schemas/Link"
                                        },
                                        "stats": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-10-29 00:00:00"
                                                    },
                                                    "hits": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Permissions error."
                    },
                    "404": {
                        "description": "Link not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Links"
                ],
                "summary": "Update all link data.",
                "operationId": "ccbf3039865493f1f7c1c7c858d825f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of link.",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutLinkRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Link updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Info message."
                                        },
                                        "entity": {
                                            "$ref": "#/components/schemas/Link"
                                        },
                                        "old_entity": {
                                            "$ref": "#/components/schemas/Link"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Permissions error."
                    },
                    "404": {
                        "description": "Link not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Links"
                ],
                "summary": "Delete link.",
                "operationId": "97e3d3c630fafdf7d97e008344c6a5f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of link.",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Link deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Permissions error."
                    },
                    "404": {
                        "description": "Link not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Links"
                ],
                "summary": "Partly update link data.",
                "operationId": "55588508e9bffa5d7aeae5c8f4a9d596",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of link.",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchLinkRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Link updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Info message."
                                        },
                                        "entity": {
                                            "$ref": "#/components/schemas/Link"
                                        },
                                        "old_entity": {
                                            "$ref": "#/components/schemas/Link"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Permissions error."
                    },
                    "404": {
                        "description": "Link not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/l/{referral}": {
            "get": {
                "tags": [
                    "Links"
                ],
                "summary": "Redirect to origin by referral.",
                "description": "Redirects to the original link, while increasing the click-through counter.",
                "operationId": "55eb6af6a2028f159053fa20ae1af272",
                "parameters": [
                    {
                        "name": "referral",
                        "in": "path",
                        "description": "Referral code.",
                        "required": true,
                        "example": "go1ZKwZNlg"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Successful redirect."
                    },
                    "404": {
                        "description": "Link not found."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "MessageResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Info message"
                    }
                },
                "type": "object"
            },
...
            "StoreLinkRequest": {
                "required": [
                    "name",
                    "description",
                    "origin"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Example group"
                    },
                    "description": {
                        "type": "string",
                        "example": "Example description"
                    },
                    "origin": {
                        "type": "string",
                        "example": "https://example.url/11111"
                    },
                    "group_id": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                },
                "type": "object"
            },
...
            "LinksListPaginated": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": "1"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Link"
                        }
                    },
                    "first_page_url": {
                        "type": "string",
                        "example": "http://links.com/links?page=1"
                    },
                    "from": {
                        "type": "integer",
                        "example": "1"
                    },
                    "last_page": {
                        "type": "integer",
                        "example": "2"
                    },
                    "last_page_url": {
                        "type": "string",
                        "example": "http://links.com/links?page=2"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "example": "null | http://links.com/links?page=2"
                                },
                                "label": {
                                    "type": "string",
                                    "example": "Next &raquo"
                                },
                                "active": {
                                    "type": "boolean",
                                    "example": "false"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "next_page_url": {
                        "type": "string",
                        "example": "null | http://links.com/links?page=2"
                    },
                    "path": {
                        "type": "string",
                        "example": "http://links.com/links"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": "10"
                    },
                    "prev_page_url": {
                        "type": "string",
                        "example": "null | http://links.com/links?page=2"
                    },
                    "to": {
                        "type": "integer",
                        "example": "10"
                    },
                    "total": {
                        "type": "integer",
                        "example": "15"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer {token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}